fix(#758): break circular dep by extracting pdf-render.ts
CI's depcruise-ratchet check caught a new circular dependency:
generate-pdf.ts dynamically imports templates/call-sheet-services.ts,
and call-sheet-services.ts statically imported renderPdfTemplate back
from generate-pdf.ts, closing the loop. Move renderPdfTemplate (the
single @pdfme/generator call site) into its own pdf-render.ts module
that neither file depends on, and update the no-restricted-imports
ignore list to match.
