+++ eslint.config.mjs
@@ -177,7 +177,7 @@ const eslintConfig = [
     // maps → @/lib/maps-sdk; Resend → src/lib/email.ts (Next) or convex/emailActions.ts
     // (Convex — two runtimes, one adapter each); PostHog → posthog-provider.tsx (client)
     // or posthog-server.ts (server) — the sanctioned error/analytics capture boundary
-    // (§8.9/§8.10); pdfme → src/lib/pdfme/generate-pdf.ts (single `generate()` call site).
+    // (§8.9/§8.10); pdfme → src/lib/pdfme/pdf-render.ts (single `generate()` call site).
     files: ["src/**/*.{ts,tsx}", "convex/**/*.ts"],
     ignores: [
       "src/lib/maps-sdk.ts",
@@ -185,7 +185,7 @@ const eslintConfig = [
       "convex/emailActions.ts",
       "src/components/providers/posthog-provider.tsx",
       "src/lib/posthog-server.ts",
-      "src/lib/pdfme/generate-pdf.ts",
+      "src/lib/pdfme/pdf-render.ts",
     ],
     rules: {
       "no-restricted-imports": [
@@ -213,7 +213,7 @@ const eslintConfig = [
             {
               name: "@pdfme/generator",
               message:
-                "Use the PDF adapter (src/lib/pdfme/generate-pdf.ts renderPdfTemplate) (R-8.10.1).",
+                "Use the PDF adapter (src/lib/pdfme/pdf-render.ts renderPdfTemplate) (R-8.10.1).",
             },
           ],
         },
